+++ Evergreen/Public/Get-1Password.ps1
@@ -43,6 +43,6 @@ Function Get-1Password {
         }
     }
     Else {
-        Write-Warning -Message "$($MyInvocation.MyCommand): unable to retreive content from $($res.Get.Update.Uri)."
+        Write-Warning -Message "$($MyInvocation.MyCommand): unable to retrieve content from $($res.Get.Update.Uri)."
     }
 }