+++ Evergreen/Public/Get-Gimp.ps1
@@ -68,7 +68,7 @@ Function Get-Gimp {
         }
     }
     Else {
-        Write-Warning -Message "$($MyInvocation.MyCommand): unable to retreive content from $($res.Get.Update.Uri)."
+        Write-Warning -Message "$($MyInvocation.MyCommand): unable to retrieve content from $($res.Get.Update.Uri)."
     }
     #endregion
 }